fix: userResult coming from DB and displaying properly
